restuct code and assets
